feat: added hosts accessor and destroyer as singleton pattern, similar to the logfile and continued work on mutli hosts parsing
Similar to how the logfile utils are implemented (static FILE *,
accessor, destroyer associated with atexit(3)), the hosts variable
(struct s_hosts *hosts) is now able to get passed around.

I could just stick it as a field of the struct s_ping structure, but
I think that's a bit counter-intuitive to using the singleton pattern..

Anyhow I think I'll fully invest in the singleton pattern, that way
we'll know its shortfalls and whatnot.