feat(docs): adding starting point for Doxygen doc
Added the main parts of a doxygen doc; the Doxyfile, a Makefile recipe
for it, also added some snippets in my config while I'm at it.

Still need some work, like documenting the rest of the structure fields,
which should take about an hour, and document all the other functions
(even static ones).

Looks sick tho in the browser